Dtype issues.

+++ losses.py
@@ -16,7 +16,7 @@ class log_loss():
                           torch.sum(torch.mul(X, w), dim=1)))))
 
     def grad(X, y, w):
-        return torch.mean(torch.div(-torch.mul(X.t(), y),
+        return torch.mean(torch.div(-torch.mul(X.t(), y.double()),
                           (1. + torch.exp(y * torch.sum(torch.mul(X, w))))),
                           dim=1)
